+++ build-template/src/frontend/vite.config.js
@@ -9,7 +9,7 @@ const ii_url =
         : `https://identity.internetcomputer.org/`;
 
 process.env.II_URL = process.env.II_URL || ii_url;
-process.env.STORAGE_GATEWAY_URL = process.env.STORAGE_GATEWAY_URL || 'https://blob.caffeine.ai';
+process.env.STORAGE_GATEWAY_URL = process.env.STORAGE_GATEWAY_URL || 'https://icp0.io';
 
 export default defineConfig({
     logLevel: 'error',